consolidate actions into single dropdown
