+++ admin/app/templates/components/post-account-modal.hbs
@@ -1,7 +1,7 @@
 {{#bs-modal show=isOpen onHidden=(action "clear") as |modal|}}
   <form {{action "updateIdentity" on="submit"}}>
     {{#modal.header}}
-      <h5 class="modal-title">Moderate User</h5>
+      <h5 class="modal-title">Modify User</h5>
       <button {{action modal.actions.hide}} type="button" class="close" aria-label="Close">
         {{entypo-icon "cross"}}
       </button>
@@ -44,15 +44,6 @@
     {{/modal.body}}
 
     {{#modal.footer}}
-    {{#if banned}}
-      <button class="btn btn-success" type="" {{action "unbanIdentity"}}>
-        {{entypo-icon "block"}} Unban User
-      </button>
-    {{else}}
-      <button class="btn btn-danger" type="" {{action "banIdentity"}}>
-        {{entypo-icon "block"}} Ban User
-      </button>
-    {{/if}}
       <button class="btn btn-success" type="submit" disabled={{isSubmitDisabled}}>
         {{entypo-icon "check"}} Update
       </button>
