+++ admin/app/templates/components/post-item-actions.hbs
@@ -1,75 +1,62 @@
-<div class="row">
-<div class="btn-group col-4" role="group">
-  {{#if item.account}}
-    <button
-      type="button"
-      class="btn btn-block btn-sm btn-outline-primary"
-      title="Moderate User"
-      {{action "moderate"}}>
-      <span class="d-none d-md-inline">Moderate user</span>
-      {{entypo-icon "user" class="ml-1"}}
-    </button>
+
+<button class="btn btn-block btn-sm btn-outline-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+  <span class="d-none d-md-inline">Moderate</span>
+  <span class="d-inline d-md-none d-lg-none">{{entypo-icon "pencil" class="ml-1"}}</span>
+</button>
+
+<div class="dropdown-menu">
+  <h6 class="dropdown-header">Post Actions</h6>
+
+  <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "edit"}}>
+    Edit {{entypo-icon "pencil" class="ml-1"}}
+  </a>
+
+  {{#if item.flagged}}
+    <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "unflag"}}>
+      Unflag {{entypo-icon "flag" class="ml-1"}}
+    </a>
   {{else}}
-    <button
-      type="button"
-      class="btn btn-block btn-sm btn-outline-disabled disabled btn-disabled"
-      title="No user present">
-      <span class="d-none d-md-inline">Moderate user</span>
-      {{entypo-icon "user" class="ml-1"}}
-    </button>
+    <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "flag"}}>
+      Flag {{entypo-icon "flag" class="ml-1"}}
+    </a>
   {{/if}}
-</div>
 
-<div class="btn-group col-4" role="group">
-{{#unless item.deleted}}
-  <button class="btn btn-block btn-sm btn-outline-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-    <span class="d-none d-md-inline">Moderate post</span>
-    <span class="d-inline d-md-none d-lg-none">{{entypo-icon "pencil" class="ml-1"}}</span>
-  </button>
-
-  <div class="dropdown-menu">
-    <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "edit"}}>
-      Edit post {{entypo-icon "pencil" class="ml-1"}}
+  {{#if item.approved}}
+    <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "unapprove"}}>
+      Hide {{entypo-icon "eye-with-line" class="ml-1"}}
     </a>
+  {{else}}
+    <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "approve"}}>
+      Show {{entypo-icon "eye" class="ml-1"}}
+    </a>
+  {{/if}}
 
-    {{#if item.flagged}}
-      <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "unflag"}}>
-        Unflag post {{entypo-icon "flag" class="ml-1"}}
-      </a>
-    {{else}}
-      <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "flag"}}>
-        Flag post {{entypo-icon "flag" class="ml-1"}}
-      </a>
-    {{/if}}
-
-    {{#if item.approved}}
-      <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "unapprove"}}>
-        Hide post {{entypo-icon "eye-with-line" class="ml-1"}}
-      </a>
-    {{else}}
-      <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "approve"}}>
-        Show post {{entypo-icon "eye" class="ml-1"}}
-      </a>
-    {{/if}}
-
+  {{#if item.deleted}}
+    <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "undelete"}}>
+      Undelete {{entypo-icon "trash" class="ml-1"}}
+    </a>
+  {{else}}
     <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "delete"}}>
-      Delete post {{entypo-icon "trash" class="ml-1"}}
+      Delete {{entypo-icon "trash" class="ml-1"}}
     </a>
-  </div>
+  {{/if}}
 
-{{else}}
-  <button class="btn btn-block btn-sm btn-outline-primary disabled" title="Post is deleted">
-    <span class="d-none d-sm-inline">
-      Moderate post {{entypo-icon "trash" class="ml-1"}}
-    </span>
-  </button>
-{{/unless}}
-</div>
-
-<div class="btn-group col-4" role="group">
-    <a href="{{ item.stream.url }}" class="btn btn-block btn-sm btn-outline-primary" target="_blank" title="View Thread">
-      {{entypo-icon "popup"}}
-      <span class="d-none d-md-inline">View Thread</span>
+  <h6 class="dropdown-header">User Actions</h6>
+  {{#if item.account}}
+    <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "moderate"}}>
+      Edit {{entypo-icon "user" class="ml-1"}}
     </a>
-</div>
+    {{#if item.account.settings.shadowbanned}}
+      <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "unban"}}>
+        Unban {{entypo-icon "check" class="ml-1"}}
+      </a>
+    {{else}}
+      <a href="javascript:void(0)" class="d-flex justify-content-between dropdown-item" {{action "ban"}}>
+        Ban {{entypo-icon "block" class="ml-1"}}
+      </a>
+    {{/if}}
+  {{else}}
+    <div class="dropdown-item text-muted">No actions available</div>
+  {{/if}}
+
 </div>
